refactor: make a big refactoring
