+++ docs/IncomingBrowserEvent/README.md
@@ -1,10 +1,10 @@
-[**Browser Core Documentation v0.0.2**](../../README.md)
+[**Browser Core Documentation v0.0.2**](../README.md)
 
 ***
 
-[Browser Core Documentation](../../modules.md) / events/IncomingBrowserEvent
+[Browser Core Documentation](../modules.md) / IncomingBrowserEvent
 
-# events/IncomingBrowserEvent
+# IncomingBrowserEvent
 
 ## Classes
