+++ docs/IncomingBrowserEvent/interfaces/IncomingBrowserEventOptions.md
@@ -1,0 +1,111 @@
+[**Browser Core Documentation v0.0.2**](../../README.md)
+
+***
+
+[Browser Core Documentation](../../modules.md) / [IncomingBrowserEvent](../README.md) / IncomingBrowserEventOptions
+
+# Interface: IncomingBrowserEventOptions
+
+Defined in: browser-core/src/IncomingBrowserEvent.ts:10
+
+IncomingBrowserEventOptions interface.
+
+## Extends
+
+- `IncomingEventOptions`
+
+## Indexable
+
+\[`key`: `string`\]: `unknown`
+
+## Properties
+
+### cookies?
+
+> `optional` **cookies**: [`CookieCollection`](../../cookies/CookieCollection/classes/CookieCollection.md)
+
+Defined in: browser-core/src/IncomingBrowserEvent.ts:14
+
+***
+
+### locale?
+
+> `optional` **locale**: `string`
+
+Defined in: core/dist/index.d.ts:148
+
+#### Inherited from
+
+`IncomingEventOptions.locale`
+
+***
+
+### metadata?
+
+> `optional` **metadata**: `Record`\<`string`, `unknown`\>
+
+Defined in: core/dist/index.d.ts:32
+
+#### Inherited from
+
+`IncomingEventOptions.metadata`
+
+***
+
+### protocol?
+
+> `optional` **protocol**: `string`
+
+Defined in: browser-core/src/IncomingBrowserEvent.ts:12
+
+***
+
+### queryString?
+
+> `optional` **queryString**: `string`
+
+Defined in: browser-core/src/IncomingBrowserEvent.ts:13
+
+***
+
+### source
+
+> **source**: `IncomingEventSource`
+
+Defined in: core/dist/index.d.ts:149
+
+#### Inherited from
+
+`IncomingEventOptions.source`
+
+***
+
+### timeStamp?
+
+> `optional` **timeStamp**: `number`
+
+Defined in: core/dist/index.d.ts:31
+
+#### Inherited from
+
+`IncomingEventOptions.timeStamp`
+
+***
+
+### type?
+
+> `optional` **type**: `string`
+
+Defined in: core/dist/index.d.ts:29
+
+#### Inherited from
+
+`IncomingEventOptions.type`
+
+***
+
+### url
+
+> **url**: `URL`
+
+Defined in: browser-core/src/IncomingBrowserEvent.ts:11
