+++ docs/OutgoingBrowserResponse/interfaces/OutgoingBrowserResponseOptions.md
@@ -1,0 +1,103 @@
+[**Browser Core Documentation v0.0.2**](../../README.md)
+
+***
+
+[Browser Core Documentation](../../modules.md) / [OutgoingBrowserResponse](../README.md) / OutgoingBrowserResponseOptions
+
+# Interface: OutgoingBrowserResponseOptions
+
+Defined in: browser-core/src/OutgoingBrowserResponse.ts:6
+
+Options for creating an Outgoing Browser Response.
+
+## Extends
+
+- `OutgoingResponseOptions`
+
+## Indexable
+
+\[`key`: `string`\]: `unknown`
+
+## Properties
+
+### content?
+
+> `optional` **content**: `unknown`
+
+Defined in: core/dist/index.d.ts:256
+
+#### Inherited from
+
+`OutgoingResponseOptions.content`
+
+***
+
+### metadata?
+
+> `optional` **metadata**: `Record`\<`string`, `unknown`\>
+
+Defined in: core/dist/index.d.ts:32
+
+#### Inherited from
+
+`OutgoingResponseOptions.metadata`
+
+***
+
+### source?
+
+> `optional` **source**: `object`
+
+Defined in: core/dist/index.d.ts:30
+
+#### Inherited from
+
+`OutgoingResponseOptions.source`
+
+***
+
+### statusCode?
+
+> `optional` **statusCode**: `number`
+
+Defined in: core/dist/index.d.ts:257
+
+#### Inherited from
+
+`OutgoingResponseOptions.statusCode`
+
+***
+
+### statusMessage?
+
+> `optional` **statusMessage**: `string`
+
+Defined in: core/dist/index.d.ts:258
+
+#### Inherited from
+
+`OutgoingResponseOptions.statusMessage`
+
+***
+
+### timeStamp?
+
+> `optional` **timeStamp**: `number`
+
+Defined in: core/dist/index.d.ts:31
+
+#### Inherited from
+
+`OutgoingResponseOptions.timeStamp`
+
+***
+
+### type?
+
+> `optional` **type**: `string`
+
+Defined in: core/dist/index.d.ts:29
+
+#### Inherited from
+
+`OutgoingResponseOptions.type`
