+++ docs/cookies/Cookie/classes/Cookie.md
@@ -6,7 +6,7 @@
 
 # Class: Cookie
 
-Defined in: [cookies/Cookie.ts:7](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L7)
+Defined in: [browser-core/src/cookies/Cookie.ts:7](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L7)
 
 Class representing a Cookie.
 
@@ -16,7 +16,7 @@ Class representing a Cookie.
 
 > `protected` **new Cookie**(`name`, `value`, `options`): [`Cookie`](Cookie.md)
 
-Defined in: [cookies/Cookie.ts:27](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L27)
+Defined in: [browser-core/src/cookies/Cookie.ts:27](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L27)
 
 #### Parameters
 
@@ -48,7 +48,7 @@ Cookie options.
 
 > `readonly` **name**: `string`
 
-Defined in: [cookies/Cookie.ts:8](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L8)
+Defined in: [browser-core/src/cookies/Cookie.ts:8](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L8)
 
 ***
 
@@ -56,7 +56,7 @@ Defined in: [cookies/Cookie.ts:8](https://github.com/stonemjs/browser-core/blob/
 
 > `readonly` **options**: [`CookieOptions`](../../../declarations/interfaces/CookieOptions.md)
 
-Defined in: [cookies/Cookie.ts:10](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L10)
+Defined in: [browser-core/src/cookies/Cookie.ts:10](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L10)
 
 ***
 
@@ -64,7 +64,7 @@ Defined in: [cookies/Cookie.ts:10](https://github.com/stonemjs/browser-core/blob
 
 > `readonly` **value**: `unknown`
 
-Defined in: [cookies/Cookie.ts:9](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L9)
+Defined in: [browser-core/src/cookies/Cookie.ts:9](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L9)
 
 ## Methods
 
@@ -72,7 +72,7 @@ Defined in: [cookies/Cookie.ts:9](https://github.com/stonemjs/browser-core/blob/
 
 > **cloneWith**(`value`, `options`): [`Cookie`](Cookie.md)
 
-Defined in: [cookies/Cookie.ts:70](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L70)
+Defined in: [browser-core/src/cookies/Cookie.ts:70](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L70)
 
 Clone the cookie with new name, value, and options.
 
@@ -102,7 +102,7 @@ A new cookie instance.
 
 > **serialize**(): `string`
 
-Defined in: [cookies/Cookie.ts:54](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L54)
+Defined in: [browser-core/src/cookies/Cookie.ts:54](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L54)
 
 Serialize the cookie value.
 
@@ -116,7 +116,7 @@ Serialize the cookie value.
 
 > **setExpires**(`value`): `this`
 
-Defined in: [cookies/Cookie.ts:37](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L37)
+Defined in: [browser-core/src/cookies/Cookie.ts:37](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L37)
 
 Set expiration date for the cookie.
 
@@ -138,7 +138,7 @@ Expiration date.
 
 > **setSecure**(`value`): `this`
 
-Defined in: [cookies/Cookie.ts:46](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L46)
+Defined in: [browser-core/src/cookies/Cookie.ts:46](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L46)
 
 Set secure flag for the cookie.
 
@@ -160,7 +160,7 @@ Whether the cookie is secure.
 
 > `static` **create**(`name`, `value`, `options`): [`Cookie`](Cookie.md)
 
-Defined in: [cookies/Cookie.ts:18](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/cookies/Cookie.ts#L18)
+Defined in: [browser-core/src/cookies/Cookie.ts:18](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/cookies/Cookie.ts#L18)
 
 Create a Cookie.
