+++ docs/declarations/enumerations/CookieSameSite.md
@@ -6,7 +6,7 @@
 
 # Enumeration: CookieSameSite
 
-Defined in: [declarations.ts:17](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/declarations.ts#L17)
+Defined in: [browser-core/src/declarations.ts:17](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/declarations.ts#L17)
 
 Enum representing possible values for the `SameSite` attribute in cookies.
 
@@ -16,7 +16,7 @@ Enum representing possible values for the `SameSite` attribute in cookies.
 
 > **Lax**: `"lax"`
 
-Defined in: [declarations.ts:18](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/declarations.ts#L18)
+Defined in: [browser-core/src/declarations.ts:18](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/declarations.ts#L18)
 
 ***
 
@@ -24,7 +24,7 @@ Defined in: [declarations.ts:18](https://github.com/stonemjs/browser-core/blob/2
 
 > **None**: `"none"`
 
-Defined in: [declarations.ts:19](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/declarations.ts#L19)
+Defined in: [browser-core/src/declarations.ts:19](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/declarations.ts#L19)
 
 ***
 
@@ -32,4 +32,4 @@ Defined in: [declarations.ts:19](https://github.com/stonemjs/browser-core/blob/2
 
 > **Strict**: `"strict"`
 
-Defined in: [declarations.ts:20](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/declarations.ts#L20)
+Defined in: [browser-core/src/declarations.ts:20](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/declarations.ts#L20)
