+++ docs/errors/BrowserError/classes/BrowserError.md
@@ -6,7 +6,7 @@
 
 # Class: BrowserError
 
-Defined in: [errors/BrowserError.ts:6](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/errors/BrowserError.ts#L6)
+Defined in: [browser-core/src/errors/BrowserError.ts:6](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/errors/BrowserError.ts#L6)
 
 Custom error for Browser operations.
 
@@ -20,7 +20,7 @@ Custom error for Browser operations.
 
 > **new BrowserError**(`message`, `options`?): [`BrowserError`](BrowserError.md)
 
-Defined in: [errors/BrowserError.ts:7](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/errors/BrowserError.ts#L7)
+Defined in: [browser-core/src/errors/BrowserError.ts:7](https://github.com/stonemjs/browser-core/blob/58b3a039142f6865ef6912ef058985d46f08d508/src/errors/BrowserError.ts#L7)
 
 #### Parameters
 
@@ -39,3 +39,103 @@ Defined in: [errors/BrowserError.ts:7](https://github.com/stonemjs/browser-core/
 #### Overrides
 
 `InitializationError.constructor`
+
+## Properties
+
+### cause?
+
+> `readonly` `optional` **cause**: `Error`
+
+Defined in: core/dist/index.d.ts:3470
+
+#### Inherited from
+
+`InitializationError.cause`
+
+***
+
+### code?
+
+> `readonly` `optional` **code**: `string`
+
+Defined in: core/dist/index.d.ts:3469
+
+#### Inherited from
+
+`InitializationError.code`
+
+***
+
+### metadata?
+
+> `readonly` `optional` **metadata**: `unknown`
+
+Defined in: core/dist/index.d.ts:3471
+
+#### Inherited from
+
+`InitializationError.metadata`
+
+## Methods
+
+### toString()
+
+> **toString**(`multiline`?): `string`
+
+Defined in: core/dist/index.d.ts:3492
+
+Converts the error to a formatted string representation.
+
+#### Parameters
+
+##### multiline?
+
+`boolean`
+
+Determine if output value must be multiline or not.
+
+#### Returns
+
+`string`
+
+A formatted error string.
+
+#### Inherited from
+
+`InitializationError.toString`
+
+***
+
+### create()
+
+> `static` **create**\<`T`\>(`message`, `options`?): `T`
+
+Defined in: core/dist/index.d.ts:3478
+
+Create a RuntimeError.
+
+#### Type Parameters
+
+• **T** *extends* `RuntimeError` = `RuntimeError`
+
+#### Parameters
+
+##### message
+
+`string`
+
+##### options?
+
+`ErrorOptions`
+
+The options to create a RuntimeError.
+
+#### Returns
+
+`T`
+
+A new RuntimeError instance.
+
+#### Inherited from
+
+`InitializationError.create`
