--- docs/events/IncomingBrowserEvent/interfaces/IncomingBrowserEventOptions.md
@@ -1,51 +1,0 @@
-[**Browser Core Documentation v0.0.2**](../../../README.md)
-
-***
-
-[Browser Core Documentation](../../../modules.md) / [events/IncomingBrowserEvent](../README.md) / IncomingBrowserEventOptions
-
-# Interface: IncomingBrowserEventOptions
-
-Defined in: [events/IncomingBrowserEvent.ts:10](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/IncomingBrowserEvent.ts#L10)
-
-IncomingBrowserEventOptions interface.
-
-## Extends
-
-- `IncomingEventOptions`
-
-## Indexable
-
-\[`key`: `string`\]: `unknown`
-
-## Properties
-
-### cookies?
-
-> `optional` **cookies**: [`CookieCollection`](../../../cookies/CookieCollection/classes/CookieCollection.md)
-
-Defined in: [events/IncomingBrowserEvent.ts:14](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/IncomingBrowserEvent.ts#L14)
-
-***
-
-### protocol?
-
-> `optional` **protocol**: `string`
-
-Defined in: [events/IncomingBrowserEvent.ts:12](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/IncomingBrowserEvent.ts#L12)
-
-***
-
-### queryString?
-
-> `optional` **queryString**: `string`
-
-Defined in: [events/IncomingBrowserEvent.ts:13](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/IncomingBrowserEvent.ts#L13)
-
-***
-
-### url
-
-> **url**: `URL`
-
-Defined in: [events/IncomingBrowserEvent.ts:11](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/IncomingBrowserEvent.ts#L11)
