--- docs/events/OutgoingBrowserResponse/interfaces/OutgoingBrowserResponseOptions.md
@@ -1,19 +1,0 @@
-[**Browser Core Documentation v0.0.2**](../../../README.md)
-
-***
-
-[Browser Core Documentation](../../../modules.md) / [events/OutgoingBrowserResponse](../README.md) / OutgoingBrowserResponseOptions
-
-# Interface: OutgoingBrowserResponseOptions
-
-Defined in: [events/OutgoingBrowserResponse.ts:6](https://github.com/stonemjs/browser-core/blob/2c2c45da7146109ea5ae39ff81ac0b60630dfeee/src/events/OutgoingBrowserResponse.ts#L6)
-
-Options for creating an Outgoing Browser Response.
-
-## Extends
-
-- `OutgoingResponseOptions`
-
-## Indexable
-
-\[`key`: `string`\]: `unknown`
