+++ src/IncomingBrowserEvent.ts
@@ -1,7 +1,7 @@
-import { Cookie } from '../cookies/Cookie'
-import { HttpMethod, IRoute } from '../declarations'
-import { BrowserError } from '../errors/BrowserError'
-import { CookieCollection } from '../cookies/CookieCollection'
+import { Cookie } from './cookies/Cookie'
+import { HttpMethod, IRoute } from './declarations'
+import { BrowserError } from './errors/BrowserError'
+import { CookieCollection } from './cookies/CookieCollection'
 import { IncomingEvent, IncomingEventOptions } from '@stone-js/core'
 
 /**
@@ -294,7 +294,7 @@ export class IncomingBrowserEvent extends IncomingEvent {
    * @param resolver - The route resolver function.
    * @returns The current instance for method chaining.
    */
-  setRouteResolver (resolver: () => IRoute): this {
+  setRouteResolver<RouteType extends IRoute = IRoute>(resolver: () => RouteType): this {
     this.routeResolver = resolver
     return this
   }
@@ -304,8 +304,8 @@ export class IncomingBrowserEvent extends IncomingEvent {
    *
    * @returns The route parameter or the route object.
    */
-  getRoute (): IRoute | undefined {
-    return this.routeResolver?.()
+  getRoute<RouteType extends IRoute = IRoute>(): RouteType | undefined {
+    return this.routeResolver?.() as (RouteType | undefined)
   }
 
   /**
